pull out event history
